requirements.txt
youtube-transcript-api
google-generativeai
googletrans==4.0.0-rc1
nltk
python-dotenv
